Moved antenna radiation patter plot to antenna_beamforming_imt.py file
When run directly, antenna_beamforming_imt.py plots the radiation pattern
